Config (Not Working) + A few mini updates
This small development update should work on 1.21 (No 1.21.1 support yet!) but could break due to many errors with in my config menu, This config menu will have a few options, If you want to fix the menu consider forking this development branch and fixing it. I do plan to put a update to the releases after I test and finish the menu, I hope to get it out soon!